+++ CHANGELOG.md
@@ -45,7 +45,7 @@
 - PR #5658 Add `filter_tokens` nvtext API
 - PR #5666 Add `filter_characters_of_type` strings API
 - PR #5673 Always build and test with per-thread default stream enabled in the GPU CI build
-- PR #5716 Add `dispatch_two_types` to libcudf
+- PR #5716 Add `type_double_dispatcher` to libcudf
 
 ## Improvements